# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: Go LineLess API specification — Ahmedabad Pilot
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/send-otp:
    post:
      operationId: sendOtp
      tags: [auth]
      summary: Send OTP to phone (user or runner)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendOtpInput"
      responses:
        "200":
          description: OTP sent (otp returned in dev)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OtpResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
  /auth/verify-otp:
    post:
      operationId: verifyOtp
      tags: [auth]
      summary: Verify OTP and login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyOtpInput"
      responses:
        "200":
          description: Authenticated — returns token + user/runner profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          $ref: "#/components/responses/ErrorResponse"
  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout (deletes session token)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /admin/login:
    post:
      operationId: adminLogin
      tags: [auth]
      summary: Admin login with admin username and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginInput"
      responses:
        "200":
          description: Admin authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          $ref: "#/components/responses/ErrorResponse"

  # ─── PRICING ────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    ErrorResponse:
      description: Error response
      content:
        application/json:
          schema:
            type: object
            required: [error]
            properties:
              error:
                type: string
              detail:
                type: string
              code:
                type: string
  schemas:
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string

    # ─── AUTH SCHEMAS ────────────────────────────────────────────────────────
    SendOtpInput:
      type: object
      required: [phone]
      properties:
        phone:
          type: string
        role:
          type: string
          enum: [user, runner]
    OtpResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
        otp:
          type: string
          description: Returned only in dev mode
        sent:
          type: boolean
          description: Whether the SMS was sent successfully
    VerifyOtpInput:
      type: object
      required: [phone, otp]
      properties:
        phone:
          type: string
        otp:
          type: string
        role:
          type: string
          enum: [user, runner]
    AuthResponse:
      type: object
      required: [token, role]
      properties:
        token:
          type: string
        role:
          type: string
        user:
          $ref: "#/components/schemas/User"
        runner:
          $ref: "#/components/schemas/Runner"
    AdminLoginInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string

    # ─── PRICING SCHEMAS ──────────────────────────────────────────────────────
...
    User:
      type: object
      required: [id, phone]
      properties:
        id:
          type: integer
        name:
          type: string
          nullable: true
        phone:
          type: string
        email:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        area:
          type: string
          nullable: true
        avatar:
          type: string
          nullable: true
        language:
          type: string
        createdAt:
          type: string
...
    Runner:
      type: object
      required: [id, phone]
      properties:
        id:
          type: integer
        name:
          type: string
          nullable: true
        fullName:
          type: string
          nullable: true
        phone:
          type: string
        email:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        area:
          type: string
          nullable: true
        avatar:
          type: string
          nullable: true
        gender:
          type: string
          nullable: true
        kycStatus:
          type: string
          enum: [pending, verified, rejected, suspended]
        kycRejectionReason:
          type: string
          nullable: true
        dispatchAllowed:
          type: boolean
        isOnline:
          type: boolean
        rating:
          type: number
          nullable: true
        trustScore:
          type: integer
        trustBadge:
          type: string
        tasksCompleted:
          type: integer
        tasksAccepted:
          type: integer
        tasksCancelled:
          type: integer
        totalTasks:
          type: integer
        totalEarnings:
          type: number
          nullable: true
        averageRating:
          type: number
          nullable: true
        averageResponseTime:
          type: number
          nullable: true
        currentLat:
          type: number
          nullable: true
        currentLng:
          type: number
          nullable: true
        gpsStatus:
          type: string
          nullable: true
        gpsCheckedAt:
          type: string
          nullable: true
        onboardingStep:
          type: integer
          nullable: true
        onboardingCompleted:
          type: boolean
        selfie:
          type: string
          nullable: true
        createdAt:
          type: string
...